Переименование интерфейсов
