+++ src/main/java/ru/promo/consul_plan/controller/AccountController.java
@@ -8,8 +8,8 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.RestController;
 import ru.promo.consul_plan.domain.entity.ClientEntity;
 import ru.promo.consul_plan.domain.entity.SpecialistEntity;
-import ru.promo.consul_plan.service.ClientService;
-import ru.promo.consul_plan.service.SpecialistService;
+import ru.promo.consul_plan.service.ClientServiceImpl;
+import ru.promo.consul_plan.service.SpecialistServiceImpl;
 
 import java.util.List;
 
@@ -18,10 +18,10 @@ import java.util.List;
 public class AccountController implements AccountApi {
 
     @Autowired
-    private SpecialistService specialistService;
+    private SpecialistServiceImpl specialistServiceImpl;
 
     @Autowired
-    private ClientService clientService;
+    private ClientServiceImpl clientServiceImpl;
 
     //Под авторизацией всем
     //    @GetMapping("/all-auth")
@@ -35,12 +35,12 @@ public class AccountController implements AccountApi {
     }
 
     public ResponseEntity<List<SpecialistEntity>> getAllSpecialists() {
-        List<SpecialistEntity> specialists = specialistService.getAll();
+        List<SpecialistEntity> specialists = specialistServiceImpl.getAll();
         return ResponseEntity.ok(specialists);
     }
 
     public ResponseEntity<List<ClientEntity>> getAllClients() {
-        List<ClientEntity> clients = clientService.getAll();
+        List<ClientEntity> clients = clientServiceImpl.getAll();
         return ResponseEntity.ok(clients);
     }
 }
