+++ src/main/java/ru/promo/consul_plan/controller/NotificationController.java
@@ -5,8 +5,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.promo.consul_plan.domain.entity.NotificationEntity;
-import ru.promo.consul_plan.service.ConsultationService;
-import ru.promo.consul_plan.service.NotificationService;
+import ru.promo.consul_plan.service.ConsultationServiceImpl;
+import ru.promo.consul_plan.service.NotificationServiceImpl;
 
 import java.util.List;
 
@@ -14,43 +14,43 @@ import java.util.List;
 public class NotificationController implements NotificationApi{
 
     @Autowired
-    private NotificationService notificationService;
+    private NotificationServiceImpl notificationServiceImpl;
     @Autowired
-    private ConsultationService consultationService;
+    private ConsultationServiceImpl consultationServiceImpl;
 
 
     public ResponseEntity<NotificationEntity> createNotification(@Valid @RequestBody NotificationEntity notification) {
-        notificationService.create(notification);
+        notificationServiceImpl.create(notification);
         return ResponseEntity.ok(notification);
     }
 
     public ResponseEntity<NotificationEntity> getNotificationById(@PathVariable(name = "id") Long id) {
-        NotificationEntity notification = notificationService.getById(id);
+        NotificationEntity notification = notificationServiceImpl.getById(id);
         return ResponseEntity.ok(notification);
     }
     public ResponseEntity<NotificationEntity> updateNotification(@Valid@RequestBody NotificationEntity notification) {
-        notificationService.update(notification);
+        notificationServiceImpl.update(notification);
         return ResponseEntity.ok(notification);
     }
 
     public ResponseEntity<Void> deleteNotification(@PathVariable(name = "id") Long id) {
-        notificationService.delete(id);
+        notificationServiceImpl.delete(id);
         return ResponseEntity.noContent().build();
     }
 
     public ResponseEntity<List<NotificationEntity>> getNotificationsByConsultationId(@PathVariable(name = "consultationId") Long consultationId) {
-        List<NotificationEntity> notificationsEntity = notificationService.getAllByConsultationId(consultationId);
+        List<NotificationEntity> notificationsEntity = notificationServiceImpl.getAllByConsultationId(consultationId);
         return ResponseEntity.ok(notificationsEntity);
     }
 
     public ResponseEntity<List<NotificationEntity>> getNotificationsByClientId(@PathVariable(name = "clientId") Long clientId) {
-        List<NotificationEntity> notificationsEntity = notificationService.getAllByClientId(clientId);
+        List<NotificationEntity> notificationsEntity = notificationServiceImpl.getAllByClientId(clientId);
         return ResponseEntity.ok(notificationsEntity);
     }
 
     public ResponseEntity<Void> sendReminder(@PathVariable(name = "consultationId") Long consultationId) {
-        var consultation = consultationService.getById(consultationId);
-        notificationService.sendReminder(consultation);
+        var consultation = consultationServiceImpl.getById(consultationId);
+        notificationServiceImpl.sendReminder(consultation);
         return ResponseEntity.noContent().build();
     }
 }
